Mojo added a totals column.  Weekend column is before this. Need to truncate to int when sending values to nerd site.

+++ MovieMiner.Tests/MineAllTests.cs
@@ -906,7 +906,7 @@ namespace MovieMiner.Tests
 					movieList += ",";
 				}
 
-				movieList += movie.Earnings.ToString();
+				movieList += ((int)movie.Earnings).ToString();
 			}
 
 			Logger.WriteLine(url + movieList);
